requirements.txt
# Exam-Architect AI
# Python 3.12+

# Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Config
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
greenlet>=3.0.0

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0
Pillow>=10.2.0
pytesseract>=0.3.10
reportlab>=4.2.0
# unstructured>=0.12.0  # optional; uncomment for more formats

# AI
openai>=1.12.0
google-generativeai>=0.8.0
# Resolve X509_V_FLAG_NOTIFY_POLICY / SSL compatibility with Google APIs
pyopenssl>=23.2.0
cryptography>=42.0.0

# Vector store
chromadb>=0.4.22

# Task queue
celery[redis]>=5.3.0
redis>=5.0.0

# Utils
python-multipart>=0.0.9
